Code cleaning, Board Specific functions, ADC decoding, memory optimization, etc...
git-svn-id: svn+ssh://svn.gel.usherbrooke.ca/grp/svn/f12eq04/Code/Controle/Code_VUE32@106 5c0d3ecc-3e0b-11e1-b269-da1b321ef365

+++ VUE32_1_0/src/offboard_sensors.h
@@ -25,8 +25,8 @@
 //                                                                         //
 /////////////////////////////////////////////////////////////////////////////
 
-short read_current(unsigned int adc_in, unsigned int adc_vbat);
-void read_yaw_lateral(unsigned int adc_yaw, unsigned int adc_lateral);
+short read_current(unsigned short adc_in, unsigned short adc_vbat);
+void read_yaw_lateral(unsigned short adc_yaw, unsigned short adc_lateral);
 unsigned char gfi_freq_sensor(void);
 
 #endif
